// src/app/shared/generated/model/agentAbstract.ts
/**
 * onecx-chat-bff
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AgentFilter } from './agentFilter';


export interface AgentAbstract { 
    id?: string;
    name?: string;
    description?: string;
    filter?: AgentFilter;
}